% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{add_time_resolution}
\alias{add_time_resolution}
\title{Determine the time resolution of the input data tibble.}
\usage{
add_time_resolution(data_tbl)
}
\arguments{
\item{data_tbl}{Data tibble for which the time resolution should be determined.}
}
\description{
This function is designed to determine the time resolution of the input data
tibble based on the first two unique time points. The determined time resolution
/(e.g. 'days'/) is then attached to the object as attribute 'time_resolution'.
}
