% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aggregation.R
\name{aggregate_time_periods}
\alias{aggregate_time_periods}
\title{Aggregate longitudinal data to another time unit}
\usage{
aggregate_time_periods(data_tbl, aggregate_to = NULL,
  aggregation_method = sum)
}
\arguments{
\item{data_tbl}{input data tibble.}

\item{aggregate_to}{unit to which the data should be aggregated to.}

\item{aggregation_method}{function which should be used to summarize
the data per time frame.}
}
\description{
This function is designed to aggregate longitudinal data to another, bigger
time unit
}
