% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shape_clustering.R
\name{cluster_shapes}
\alias{cluster_shapes}
\title{Cluster data according to shape via kmlShape algorithm.}
\usage{
cluster_shapes(data_tbl, shapes = 3,
  n_senators = senators_function(data_tbl),
  n_times = n_times_function(data_tbl), timescale_factor = 0.1,
  frechet_method = "sum")
}
\arguments{
\item{data_tbl}{data tibble which the clustering should be performed on.}

\item{shapes}{number of shape clusters.}

\item{n_senators}{number of senators used for complexity reduction.}

\item{n_times}{number of curve description points used for complexity reduction.}

\item{timescale_factor}{factor anywhere between 0 (DTW) and 1 (euclidean distance).}

\item{frechet_method}{method used to compute Frechet distance ('sum' or 'max').}
}
\description{
This function is designed to cluster longitudinal data. The clustering
algorithm implemented in the kmlShape and used here uses es shape-respecting
clustering distance (Frechet distance) and a shape-respecting mean.
}
