% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{create_cluster_hist}
\alias{create_cluster_hist}
\title{Display number of data chunks per cluster in an interactive histogram.}
\usage{
create_cluster_hist(clustering_result)
}
\arguments{
\item{clustering_result}{Clustering result list from cluster_shapes function.}
}
\description{
This functions displays the number of unique IDs per cluster. The IDs represent
e.g. weeks, years, core periods or all data of own patient.
}
