% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{create_cluster_values_heatmap}
\alias{create_cluster_values_heatmap}
\title{Create Heatmap which displays values over time per data chunk and cluster.}
\usage{
create_cluster_values_heatmap(data_tbl, clustering_result = NULL,
  color_palette = c(12, "Paired"), ...)
}
\arguments{
\item{data_tbl}{Data tibble where the clustering was performed on.}

\item{clustering_result}{Clustering result list from cluster_shapes function.}

\item{color_palette}{Customizable color palette vector with ColorBrewer parameters.}

\item{...}{Any additional plotly arguments for ggplot2.}
}
\description{
This function creates one heatmap per cluster which shows the values per
cluster for each individual data chunk over time.
}
