% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{create_cluster_violin_plot}
\alias{create_cluster_violin_plot}
\title{Create violin plot per cluster.}
\usage{
create_cluster_violin_plot(data_tbl, additional_data_tbl,
  clustering_result = NULL, color_palette = c(12, "Paired"), ...)
}
\arguments{
\item{data_tbl}{Data tibble where the clustering was performed on.}

\item{additional_data_tbl}{Data tibble where the clustering was performed on
or any other data tibble with additional values but the same IDs and similar
time points.}

\item{clustering_result}{Clustering result list from cluster_shapes function.}

\item{color_palette}{Customizable color palette vector with ColorBrewer parameters.}

\item{...}{Any additional plotly arguments for ggplot2.}
}
\description{
This function creates violin plots which show the distribution of values
either from the data where the clustering was performed on or some additional
data with the same identifiers. One violin plot gets created for every cluster
summarized in one figure.
}
