% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunk_extraction.R
\name{extract_days}
\alias{extract_days}
\title{Extract days from data tibble.}
\usage{
extract_days(data_tbl, partial_day = TRUE, percentage_NA = 0,
  impute_NA = FALSE, window_size = 3, ...)
}
\arguments{
\item{data_tbl}{data tibble.}

\item{partial_day}{If TRUE incomplete days are also included in the result.
This means that also days which do not have at least one data point in the
first and the last hour of the day are included in the result.}

\item{percentage_NA}{if set to 0.1 -> 10\% missing values are allowed per day.}

\item{impute_NA}{If TRUE NAs will be imputed based on weighted mean in a
certain window.}

\item{window_size}{values on both sides of the value which should be imputed
(e.g. window_size =3 gives a 7 days window).}

\item{...}{Any additional argument.}
}
\description{
This function extracts one data chunk per day.
}
