% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunk_extraction.R
\name{extract_minutes}
\alias{extract_minutes}
\title{Extract minutes from data tibble.}
\usage{
extract_minutes(data_tbl, partial_minute = TRUE, percentage_NA = 0,
  impute_NA = FALSE, window_size = 3, ...)
}
\arguments{
\item{data_tbl}{data tibble.}

\item{partial_minute}{If TRUE incomplete minutes are also included in the result.
This means that also minutes which do not have at least one data point in the
first and the last second of the minute are included in the result.}

\item{percentage_NA}{If set to 0.1 -> 10\% missing values are allowed per minute.}

\item{impute_NA}{If TRUE NAs will be imputed based on weighted mean in a
certain window.}

\item{window_size}{Values on both sides of the value which should be imputed
(e.g. window_size =3 gives a 7 data point window).}

\item{...}{Any additional argument.}
}
\description{
This function extracts one data chunk per minute.
}
