% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunk_extraction.R
\name{extract_months}
\alias{extract_months}
\title{Extract months from data tibble.}
\usage{
extract_months(data_tbl, partial_month = TRUE, percentage_NA = 0,
  impute_NA = FALSE, window_size = 3, ...)
}
\arguments{
\item{data_tbl}{data tibble.}

\item{partial_month}{If TRUE incomplete months are also included in the result.
This means that also months which do not have at least one data point in the
first and the last day of the month are included in the result.}

\item{percentage_NA}{If set to 0.1 -> 10\% missing values are allowed per month.}

\item{impute_NA}{If TRUE NAs will be imputed based on weighted mean in a
certain window.}

\item{window_size}{Number of values on both sides of the value which should be imputed
(e.g. window_size =3 gives a 7 months window).}

\item{...}{Any additional argument.}
}
\description{
This function extracts one data chunk per month.
}
