% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunk_extraction.R
\name{extract_weeks}
\alias{extract_weeks}
\title{Extract weeks from tibble.}
\usage{
extract_weeks(data_tbl, partial_weeks = FALSE, percentage_NA = 0,
  start_monday = TRUE, impute_NA = FALSE, window_size = 3, ...)
}
\arguments{
\item{data_tbl}{data tibble}

\item{partial_weeks}{if TRUE incomplete weeks are also included in the result}

\item{percentage_NA}{if set to 0.1 -> 10\% missing values are allowed per week}

\item{start_monday}{if TRUE only weeks starting on Monday will be in the output}

\item{impute_NA}{If TRUE NAs will be imputed based on weighted mean in a
certain window.}

\item{window_size}{values on both sides of the value which should be imputed
(e.g. window_size =3 gives a 7 days window).}

\item{...}{Any additional argument}
}
\description{
This function extracts week long data chunks starting with Monday as first
day of the week.
}
