% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{generate_day_data}
\alias{generate_day_data}
\title{Create example data at a daily level.}
\usage{
generate_day_data(percentage_NA = 0.2)
}
\arguments{
\item{percentage_NA}{Percentage of missing valuess (NAs) in example data.}
}
\description{
This function is used to generate mobile health sensor data as an example
data set for the use in functions of the mHealthieR package.
}
