% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shape_clustering.R
\name{n_times_function}
\alias{n_times_function}
\title{Determine the right number of curve description points for shape clustering.}
\usage{
n_times_function(data_tbl)
}
\arguments{
\item{data_tbl}{data tibble which the clustering should be performed on.}
}
\description{
This function is designed to calculate the right number of points which
describes for the the shape for complexity reduction performed by the
kmlShape reduceTraj() based on the input data.
}
