% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_cluster_data}
\alias{plot_cluster_data}
\title{Plot data per cluster in separate plot incl. mean shape of the cluster.}
\usage{
plot_cluster_data(clustering_result = NULL, data_tbl = NULL,
  color_palette = c(12, "Paired"), ...)
}
\arguments{
\item{clustering_result}{Clustering result list from cluster_shapes function.}

\item{data_tbl}{data tibble which the clustering was performed on.}

\item{color_palette}{Specify the colorbrewer palette used for the cluster.
display in the following style c(number of colors in the palette, palette name).}

\item{...}{Any additional argument.}
}
\description{
This function is designed to plot the actual data for every cluster plus
the mean shape of the cluster.
}
