% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_cluster_grid}
\alias{plot_cluster_grid}
\title{Plot join single cluster plots into one grid.}
\usage{
plot_cluster_grid(cluster_plots_list, clustering_result, color_palette = c(12,
  "Paired"), ...)
}
\arguments{
\item{cluster_plots_list}{List with plots from plot_cluster_data.}

\item{clustering_result}{Clustering result list from cluster_shapes function.}

\item{color_palette}{Specify the colorbrewer palette used for the cluster.
display in the following style c(number of colors in the palette, palette name).}

\item{...}{Any additional argument.}
}
\description{
This function is designed to take the actual data for every cluster plus
the mean shape of the cluster which are already plotted and join these
individuals plots into one grid.
}
