% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_core}
\alias{plot_core}
\title{Plot cores over time.}
\usage{
plot_core(core_tb, ...)
}
\arguments{
\item{core_tb}{tibble with cores in wide format which should be plotted}

\item{...}{Any additional argument}
}
\description{
This function plots all core period over time in a binary fashion. 1 stands
for a value at a given day and 0 for a missing value.
}
