% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_mean_shapes}
\alias{plot_mean_shapes}
\title{Display mean shape summary.}
\usage{
plot_mean_shapes(clustering_result = NULL, color_palette = c(12, "Paired"),
  yaxis_unit = NULL, ...)
}
\arguments{
\item{clustering_result}{Clustering result list from cluster_shapes function.}

\item{color_palette}{Specify the colorbrewer palette used for the cluster.
display in the following style c(number of colors in the palette, palette name).}

\item{yaxis_unit}{Unit of values displayed on the y axis over time.}

\item{...}{Any additional argument.}
}
\description{
This function is designed to display all mean shapes (every cluster) in a
summary plot over time.
}
