% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_shape_clustering.R
\name{senators_function}
\alias{senators_function}
\title{Determine the right number of senators for shape clustering.}
\usage{
senators_function(data_tbl)
}
\arguments{
\item{data_tbl}{data tibble which the clustering should be performed on.}
}
\description{
This function is designed to calculate the right number of senators for the
complexity reduction performed by the kmlShape reduceTraj() based on the
input data.
}
