% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunk_extraction.R
\name{extract_weeks}
\alias{extract_weeks}
\title{Extract weeks from tibble.}
\usage{
extract_weeks(tb, partial_weeks = FALSE, percentage_NA = 0,
  start_monday = TRUE, ...)
}
\arguments{
\item{tb}{data tibble}

\item{partial_weeks}{if TRUE incomplete weeks are also included in the result}

\item{percentage_NA}{if set to 0.1 -> 10\% missing values are allowed per week}

\item{start_monday}{if TRUE only weeks starting on Monday will be in the output}

\item{...}{Any additional argument}
}
\description{
This function extracts week long data chunks starting with Monday as first
day of the week.
}
