% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunk_extraction.R
\name{extract_weeks}
\alias{extract_weeks}
\title{Extract weeks from tibble.}
\usage{
extract_weeks(tb, part_week = FALSE, weekNA = 0, ...)
}
\arguments{
\item{tb}{data tibble}

\item{part_week}{if TRUE incomplete weeks are also included in the result}

\item{weekNA}{if set to 1 -> 1 missing value is allowed per week}

\item{...}{Any additional argument}
}
\description{
This function extracts week long data chunks starting with Monday as first
day of the week.
}
