% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunk_extraction.R
\name{extract_years}
\alias{extract_years}
\title{Extract year from data tibble.}
\usage{
extract_years(data_tbl, partial_year = FALSE, percentage_NA = 0, ...)
}
\arguments{
\item{data_tbl}{data tibble}

\item{partial_year}{if TRUE incomplete years are also included in the result}

\item{percentage_NA}{if set to 0.1 -> 10\% missing values are allowed per year}

\item{...}{Any additional argument}
}
\description{
This function extracts data chunks per year.
}
